+++ core/src/main/java/org/apache/calcite/plan/RelOptUtil.java
@@ -667,8 +667,8 @@ public abstract class RelOptUtil {
         || logic == RelOptUtil.Logic.TRUE_FALSE_UNKNOWN;
     if (!outerJoin) {
       final LogicalAggregate aggregate =
-          LogicalAggregate.create(ret, ImmutableBitSet.range(keyCount), null,
-              ImmutableList.of());
+          LogicalAggregate.create(ret, ImmutableList.of(), ImmutableBitSet.range(keyCount),
+              null, ImmutableList.of());
       return new Exists(aggregate, false, false);
     }
 
@@ -867,14 +867,17 @@ public abstract class RelOptUtil {
     final RexBuilder rexBuilder = rel.getCluster().getRexBuilder();
     List<RexNode> castExps;
     RelNode input;
+    List<RelHint> hints = ImmutableList.of();
     if (rel instanceof Project) {
       // No need to create another project node if the rel
       // is already a project.
+      final Project project = (Project) rel;
       castExps = RexUtil.generateCastExpressions(
           rexBuilder,
           castRowType,
           ((Project) rel).getProjects());
       input = rel.getInput(0);
+      hints = project.getHints();
     } else {
       castExps = RexUtil.generateCastExpressions(
           rexBuilder,
@@ -884,11 +887,11 @@ public abstract class RelOptUtil {
     }
     if (rename) {
       // Use names and types from castRowType.
-      return projectFactory.createProject(input, castExps,
+      return projectFactory.createProject(input, hints, castExps,
           castRowType.getFieldNames());
     } else {
       // Use names from rowType, types from castRowType.
-      return projectFactory.createProject(input, castExps,
+      return projectFactory.createProject(input, hints, castExps,
           rowType.getFieldNames());
     }
   }
@@ -927,13 +930,15 @@ public abstract class RelOptUtil {
               null));
     }
 
-    return LogicalAggregate.create(rel, ImmutableBitSet.of(), null, aggCalls);
+    return LogicalAggregate.create(rel, ImmutableList.of(), ImmutableBitSet.of(),
+        null, aggCalls);
   }
 
   /** @deprecated Use {@link RelBuilder#distinct()}. */
   @Deprecated // to be removed before 2.0
   public static RelNode createDistinctRel(RelNode rel) {
     return LogicalAggregate.create(rel,
+        ImmutableList.of(),
         ImmutableBitSet.range(rel.getRowType().getFieldCount()), null,
         ImmutableList.of());
   }
@@ -3381,7 +3386,7 @@ public abstract class RelOptUtil {
               : fieldNames.get(i));
       exprList.add(rexBuilder.makeInputRef(rel, source));
     }
-    return projectFactory.createProject(rel, exprList, outputNameList);
+    return projectFactory.createProject(rel, ImmutableList.of(), exprList, outputNameList);
   }
 
   /** Predicate for whether a {@link Calc} contains multisets or windowed